TIME :: move GetTime() to toolkit folder
