+++ samples/fdga/DGATask.cpp
@@ -62,7 +62,7 @@ void DGATask::operator()() {
 
             if (GetCacheResult(hash, certitude)) {
                 if (is_log && (certitude>=_threshold)){
-                    _logs += R"({"evt_id": ")" + Evt_idToString() + R"(", "time": ")" + GetTime() + "\", \"domain\": \""+ domain + "\", \"dga_prob\": " + std::to_string(certitude) + "}\n";
+                    _logs += R"({"evt_id": ")" + Evt_idToString() + R"(", "time": ")" + darwin::time_utils::GetTime() + "\", \"domain\": \""+ domain + "\", \"dga_prob\": " + std::to_string(certitude) + "}\n";
                 }
                 _certitudes.push_back(certitude);
                 DARWIN_LOG_DEBUG("DGATask:: processed entry in "
@@ -73,7 +73,7 @@ void DGATask::operator()() {
 
         certitude = Predict();
         if (is_log && (certitude>=_threshold)){
-            _logs += R"({"evt_id": ")" + Evt_idToString() + R"(", "time": ")" + GetTime() + "\", \"domain\": \""+ domain + "\", \"dga_prob\": "+std::to_string(certitude) + "}\n";
+            _logs += R"({"evt_id": ")" + Evt_idToString() + R"(", "time": ")" + darwin::time_utils::GetTime() + "\", \"domain\": \""+ domain + "\", \"dga_prob\": "+std::to_string(certitude) + "}\n";
         }
         _certitudes.push_back(certitude);
         if (_is_cache) {
@@ -85,20 +85,6 @@ void DGATask::operator()() {
 
     Workflow();
     _domains = std::vector<std::string>();
-}
-
-std::string DGATask::GetTime(){
-    char str_time[256];
-    time_t rawtime;
-    struct tm * timeinfo;
-    std::string res;
-
-    time(&rawtime);
-    timeinfo = localtime(&rawtime);
-    strftime(str_time, sizeof(str_time), "%F%Z%T%z", timeinfo);
-    res = str_time;
-
-    return res;
 }
 
 void DGATask::Workflow(){
