+++ toolkit/Time.hpp
@@ -1,0 +1,16 @@
+/// \file     Time.hpp
+/// \authors  tbertin
+/// \version  1.0
+/// \date     04/09/19
+/// \license  GPLv3
+/// \brief    Copyright (c) 2019 Advens. All rights reserved.
+
+#include <string>
+
+/// \namespace darwin
+namespace darwin {
+    /// \namespace validator
+    namespace time_utils {
+        std::string GetTime();
+    }
+}
